rm casting

+++ InfoSec/p2/p2_1/p2_1.cpp
@@ -33,7 +33,7 @@ int main()
     
     while(iout>>c)
     {
-        img << (int8_t)c;
+        img << c;
     }
 
     img.close();
